extract ParentViewHolder from Parent model
